LPS-171473 Simplifying module layout
