+++ modules/apps/apibuilder/apibuilder-api/src/main/resources/com/liferay/apibuilder/resource/packageinfo
