+++ modules/apps/apibuilder/apibuilder-impl/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay APIBuilder Implementation
+Bundle-SymbolicName: com.liferay.apibuilder.impl
+Bundle-Version: 1.0.0
