+++ modules/apps/apibuilder/apibuilder-impl/src/main/java/com/liferay/apibuilder/internal/handler/GetByPrimaryKeyOperationHandler.java
