+++ modules/apps/apibuilder/apibuilder-impl/src/main/java/com/liferay/apibuilder/internal/jaxrs/application/APIBuilderApplication.java
@@ -1,0 +1,37 @@
+/**
+ * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
+ *
+ * This library is free software; you can redistribute it and/or modify it under
+ * the terms of the GNU Lesser General Public License as published by the Free
+ * Software Foundation; either version 2.1 of the License, or (at your option)
+ * any later version.
+ *
+ * This library is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
+ */
+
+package com.liferay.apibuilder.internal.jaxrs.application;
+
+import com.liferay.apibuilder.constants.APIBuilderConstants;
+
+import javax.ws.rs.core.Application;
+
+import org.osgi.service.component.annotations.Component;
+
+/**
+ * @author Matija Petanjek
+ * @author Carlos Correa
+ */
+@Component(
+	property = {
+		"osgi.jaxrs.application.base=" + APIBuilderConstants.BASE_PATH,
+		"osgi.jaxrs.extension.select=(osgi.jaxrs.name=Liferay.APIBuilder.Feature)",
+		"osgi.jaxrs.extension.select=(osgi.jaxrs.name=Liferay.Vulcan)",
+		"osgi.jaxrs.name=Liferay.APIBuilder.Application"
+	},
+	service = Application.class
+)
+public class APIBuilderApplication extends Application {
+}
