--- modules/apps/apibuilder/apibuilder-rest-impl/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay APIBuilder REST Implementation
-Bundle-SymbolicName: com.liferay.apibuilder.rest.impl
-Bundle-Version: 1.0.0
