--- modules/apps/apibuilder/apibuilder-rest-impl/src/main/java/com/liferay/apibuilder/rest/internal/resource/BaseAPIBuilderResourceImpl.java
@@ -1,48 +1,0 @@
-/**
- * Copyright (c) 2000-present Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.apibuilder.rest.internal.resource;
-
-import com.liferay.apibuilder.rest.resource.APIBuilderResource;
-
-import javax.servlet.http.HttpServletRequest;
-
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Response;
-
-/**
- * @author Carlos Correa
- */
-public abstract class BaseAPIBuilderResourceImpl implements APIBuilderResource {
-
-	@GET
-	@Path("{any: .*}")
-	@Produces({"application/json", "application/xml"})
-	public abstract Response get() throws Exception;
-
-	public HttpServletRequest getContextHttpServletRequest() {
-		return contextHttpServletRequest;
-	}
-
-	public void setContextHttpServletRequest(
-		HttpServletRequest contextHttpServletRequest) {
-
-		this.contextHttpServletRequest = contextHttpServletRequest;
-	}
-
-	protected HttpServletRequest contextHttpServletRequest;
-
-}
