--- modules/apps/apibuilder/apibuilder-rest-test/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay APIBuilder REST Test
-Bundle-SymbolicName: com.liferay.apibuilder.rest.test
-Bundle-Version: 1.0.0
