--- modules/apps/apibuilder/apibuilder-rest-test/build.gradle
@@ -1,10 +1,0 @@
-dependencies {
-	testIntegrationCompile group: "com.liferay.portal", name: "com.liferay.portal.impl", version: "default"
-	testIntegrationCompile group: "org.apache.felix", name: "org.apache.felix.http.servlet-api", version: "1.1.2"
-	testIntegrationCompile project(":apps:apibuilder:apibuilder-api")
-	testIntegrationCompile project(":apps:apibuilder:apibuilder-rest-api")
-	testIntegrationCompile project(":apps:blogs:blogs-api")
-	testIntegrationCompile project(":apps:blogs:blogs-test-util")
-	testIntegrationCompile project(":apps:portal-vulcan:portal-vulcan-api")
-	testIntegrationCompile project(":test:arquillian-extension-junit-bridge")
-}
