--- modules/apps/apibuilder/apibuilder-service/bnd.bnd
@@ -1,3 +1,0 @@
-Bundle-Name: Liferay APIBuilder Service
-Bundle-SymbolicName: com.liferay.apibuilder.service
-Bundle-Version: 1.0.0
