+++ modules/apps/apibuilder/apibuilder-test/bnd.bnd
@@ -1,0 +1,3 @@
+Bundle-Name: Liferay APIBuilder Test
+Bundle-SymbolicName: com.liferay.apibuilder.test
+Bundle-Version: 1.0.0
