+++ modules/apps/apibuilder/apibuilder-test/src/testIntegration/resources/blogs-entry-openapi.yaml
